requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
apache-tvm-ffi==0.1.8.post2
astor==0.8.1
attrs==25.4.0
bitsandbytes==0.49.1
blake3==1.0.8
blinker==1.9.0
boto3==1.42.34
botocore==1.42.34
brotli==1.2.0
cachetools==6.2.4
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.13.0
contourpy==1.3.3
cryptography==46.0.4
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==3.6.0
datasketch==1.6.4
depyf==0.20.0
diffusers==0.36.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker-pycreds==0.4.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.3
flashinfer-python==0.6.1
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.76.0
grpcio-reflection==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
ijson==3.4.0.post0
imagesize==1.4.1
importlib_metadata==8.7.1
interegular==0.3.3
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.22.0
matplotlib==3.10.0
mcp==1.26.0
mdurl==0.1.2
megatron-core==0.15.2
mistral_common==1.9.0
ml_dtypes==0.5.4
model-hosting-container-standards==0.1.13
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.15.0
networkx==3.6.1
ngrok==1.4.0
ninja==1.13.0
nltk==3.8
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.5
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.6.77
onnx==1.20.1
onnx-ir==0.1.15
onnxscript==0.6.0
openai==2.16.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.90
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
peft==0.17.0
pillow==11.3.0
pillow-jxl-plugin==1.3.7
platformdirs==4.5.1
prettytable==3.17.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==23.0.0
pybase64==1.4.3
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyecharts==2.0.9
Pygments==2.19.2
PyJWT==2.11.0
pynvml==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==13.7.1
rich-toolkit==0.18.1
rignore==0.7.6
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.17.0
sdnq==0.1.4
sentence-transformers==2.3.1
sentencepiece==0.2.1
sentry-sdk==2.50.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
simhash==2.1.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.50.0
streamlit==1.50.0
supervisor==4.3.0
swankit==0.2.4
swanlab==0.7.6
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0+cu126
torchao==0.15.0
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.5.4
tqdm==4.67.1
transformer_engine==2.11.0
transformer_engine_cu12==2.11.0
transformer_engine_torch==2.11.0
transformers==4.57.1
triton==3.4.0
trl==0.13.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.1.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.15.0
wandb==0.18.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.3.2
websockets==16.0
Werkzeug==3.1.5
wheel==0.45.1
wrapt==2.0.1
xgrammar==0.1.29
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
